Weather bug Fixes
