--- src/functions/os_ops.py
@@ -1,28 +1,0 @@
-import os
-import subprocess as sp
-
-paths = {
-    'notepad': "C:\\Program Files\\Notepad++\\notepad++.exe",
-    'discord': "C:\\Users\\ashut\\AppData\\Local\\Discord\\app-1.0.9003\\Discord.exe",
-    'calculator': "C:\\Windows\\System32\\calc.exe"
-}
-
-
-def open_notepad():
-    os.startfile(paths['notepad'])
-
-
-def open_discord():
-    os.startfile(paths['discord'])
-
-
-def open_cmd():
-    os.system('start cmd')
-
-
-def open_camera():
-    sp.run('start microsoft.windows.camera:', shell=True)
-
-
-def open_calculator():
-    sp.Popen(paths['calculator'])
